Move metrics enum and implementations to new source file.
